more more ctf grind, Red Bot 3 is too good so i wasn't able to see if this code works.
jsut trying to have a function that gets red Bot 5 to cover the flag while.... while writing this ive realised that the scenario where bot 5 will be able to cover the flag while being out of jail and red bot 2 being in jail will never happen, since if red2 is in jail, red 5 wont be out of jail since it will always be tagged first over red2. so no matter what, if coreys bot goes down, then its either risk the flag so we can grab our bots, or stick with defense and hope for the best.